requirements.txt
pyqt5
qscintilla
pycryptodome
validators
sqlalchemy
binaryornot
requests
